requirements.macos.txt
absl-py==2.3.0
accelerate==1.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alias-free-torch==0.0.6
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argbind==0.3.9
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
auraloss==0.4.0
braceexpand==0.1.7
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
clean-fid==0.1.35
click==8.2.1
clip-anytorch==2.6.0
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
dctorch==0.1.2
decorator==5.2.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.1
einops-exts==0.0.4
einx==0.3.0
ema-pytorch==0.2.3
encodec==0.1.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.13
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.0
flatten-dict==0.4.2
fonttools==4.58.4
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
gin-config==0.5.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.34.2
gradio_client==1.10.3
groovy==0.1.2
grpcio==1.73.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
imageio==2.37.0
importlib-resources==5.12.0
inquirerpy==0.3.4
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonmerge==1.9.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
julius==0.2.7
k-diffusion==0.1.1
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
laion-clap==1.1.4
lazy_loader==0.4
librosa==0.11.0
lightning-utilities==0.14.3
llvmlite==0.43.0
local-attention==1.8.6
Markdown==3.8.2
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.0
networkx<3.3
numba==0.60.0
numpy==1.23.5
orjson==3.10.18
packaging==25.0
pandas==2.0.2
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
pfzy==0.3.4
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
prefigure==0.0.9
progressbar==2.5
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.19.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyloudnorm==0.1.1
pyparsing==3.2.3
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.1.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
randomname==0.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
ruff==0.12.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.30.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
stable-audio-tools==0.0.19
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.21.1
tomlkit==0.13.3
torch-stoi==0.2.3
torchdiffeq==0.2.5
torchlibrosa==0.1.0
torchmetrics==0.11.4
torchsde==0.2.6
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.52.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
v-diffusion-pytorch==0.0.2
vector-quantize-pytorch==1.14.41
wandb==0.15.4
wcwidth==0.2.13
webdataset==0.2.100
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
yarl==1.20.1
